DjangoFrontend/requirements.txt
django
djangorestframework
py-eureka-client